float to double conversion

+++ ar_model/ar_class.cpp
@@ -30,18 +30,18 @@ void ar_model::sample() const {
     //usage of sampler: Eigen::EigenMultivariateNormal<float> normal_sampler(mean_X, covar_X)
     //                  normal_sampler.samples(n) n no. of samples
     //Identity matrices
-    Eigen::MatrixXf id_p = Eigen::VectorXf::Ones(this->p).asDiagonal();
-    Eigen::MatrixXf id_N = Eigen::VectorXf::Ones(this->N).asDiagonal();
+    Eigen::MatrixXd id_p = Eigen::VectorXd::Ones(this->p).asDiagonal();
+    Eigen::MatrixXd id_N = Eigen::VectorXd::Ones(this->N).asDiagonal();
 
 
     //initialization
-    Eigen::VectorXf beta(this->p);
-    std::vector<Eigen::VectorXf> o_store(this->T+1);
-    Eigen::VectorXf mu_0(this->N);
-    float rho;
-    float sigma_eps = 1; //fixed for now
-    float sigma_w = 1;     // fixed for now
-    float sigma_0 = 1;      //fixed for now
+    Eigen::VectorXd beta(this->p);
+    std::vector<Eigen::VectorXd> o_store(this->T+1);
+    Eigen::VectorXd mu_0(this->N);
+    double rho;
+    double sigma_eps = 1.; //fixed for now
+    double sigma_w = 1.;     // fixed for now
+    double sigma_0 = 1.;      //fixed for now
 
 
 
@@ -49,17 +49,17 @@ void ar_model::sample() const {
     //initialization
     {
         //beta
-        Eigen::VectorXf beta_prior = Eigen::VectorXf::Zero(this->p);
-        Eigen::EigenMultivariateNormal<float> beta_sampler(beta_prior, this->beta_sig_prior*id_p);
+        Eigen::VectorXd beta_prior = Eigen::VectorXd::Zero(this->p);
+        Eigen::EigenMultivariateNormal<double> beta_sampler(beta_prior, this->beta_sig_prior*id_p);
         beta = beta_sampler.samples(1);
 
         //o vector
-        Eigen::VectorXf o_zero_prior = Eigen::VectorXf::Zero(this->N);
-        Eigen::EigenMultivariateNormal<float> o_zero_sampler(o_zero_prior, sigma_0*id_N);
+        Eigen::VectorXd o_zero_prior = Eigen::VectorXd::Zero(this->N);
+        Eigen::EigenMultivariateNormal<double> o_zero_sampler(o_zero_prior, sigma_0*id_N);
         o_store[0] = o_zero_sampler.samples(1);
         //initialize all spatial effects to zero at beginning except for index 0
         for (int t = 1; t <= this->T; ++t) {
-            o_store[t] = Eigen::VectorXf::Zero(N);
+            o_store[t] = Eigen::VectorXd::Zero(N);
         }
 
         // rho
@@ -68,9 +68,9 @@ void ar_model::sample() const {
 
         //mu_0
         //beta
-        Eigen::VectorXf mu_0_prior = Eigen::VectorXf::Zero(this->N);
-        Eigen::EigenMultivariateNormal<float> mu_0_sampler(mu_0_prior, this->mu0_sig_prior*id_N);
-        Eigen::VectorXf mu_0 = mu_0_sampler.samples(1);
+        Eigen::VectorXd mu_0_prior = Eigen::VectorXd::Zero(this->N);
+        Eigen::EigenMultivariateNormal<double> mu_0_sampler(mu_0_prior, this->mu0_sig_prior*id_N);
+        Eigen::VectorXd mu_0 = mu_0_sampler.samples(1);
 
     }
 
@@ -78,9 +78,8 @@ void ar_model::sample() const {
     // for n_iter loop: update o_store, then update beta, then update rho. write in proto files results after iteration
     for (int i = 0; i < this->n_iter; ++i) {
 
-
         //update covar inverse
-        Eigen::MatrixXf w_full_cov_inv = (1/sigma_w) * this->matern_inv;
+        Eigen::MatrixXd w_full_cov_inv = (1/sigma_w) * this->matern_inv;
 
 
 
@@ -88,11 +87,11 @@ void ar_model::sample() const {
         //o_store update
 
         //update 0 zero first
-        Eigen::MatrixXf o_zero_update_cov = post::calc_cov_eff_0(w_full_cov_inv,this->matern_inv, rho, sigma_0); // insert calculations from posterior
-        Eigen::VectorXf o_zero_update_mean = o_zero_update_cov
+        Eigen::MatrixXd o_zero_update_cov = post::calc_cov_eff_0(w_full_cov_inv,this->matern_inv, rho, sigma_0); // insert calculations from posterior
+        Eigen::VectorXd o_zero_update_mean = o_zero_update_cov
                                             * post::calc_mean_eff_0((this->X)[0], w_full_cov_inv, o_store[1],
                                                                     beta, matern_inv, mu_0, rho, sigma_0); // insert calculations from posterior
-        Eigen::EigenMultivariateNormal<float> o_zero_sampler(o_zero_update_mean, o_zero_update_cov);
+        Eigen::EigenMultivariateNormal<double> o_zero_sampler(o_zero_update_mean, o_zero_update_cov);
         o_store[0] = o_zero_sampler.samples(1);
 
         // write in proto file
@@ -107,11 +106,11 @@ void ar_model::sample() const {
         // now update the rest
 
         for (int t = 1; t < this->T-1; ++t) {
-            Eigen::MatrixXf ot_update_cov = post::calc_cov_eff_t(sigma_eps,w_full_cov_inv );
-            Eigen::VectorXf ot_update_mean = ot_update_cov
+            Eigen::MatrixXd ot_update_cov = post::calc_cov_eff_t(sigma_eps,w_full_cov_inv );
+            Eigen::VectorXd ot_update_mean = ot_update_cov
                                                 * post::calc_mean_eff_t((this->y)[t-1],(this->X)[t-1], (this->X)[t],
                                                                         w_full_cov_inv, o_store[t-1], o_store[t+1], beta, rho, sigma_eps);
-            Eigen::EigenMultivariateNormal<float> ot_sampler(ot_update_mean, ot_update_cov);
+            Eigen::EigenMultivariateNormal<double> ot_sampler(ot_update_mean, ot_update_cov);
             o_store[t] = ot_sampler.samples(1);
             //write in proto file
             o_data::vector* o_t_vector = o_matrix->add_vec(); //maybe not neccessary
@@ -120,11 +119,11 @@ void ar_model::sample() const {
         }
 
         //update last in o_store
-        Eigen::MatrixXf oT_update_cov = post::calc_cov_eff_t(sigma_eps,w_full_cov_inv );;
-        Eigen::VectorXf oT_update_mean =    oT_update_cov
+        Eigen::MatrixXd oT_update_cov = post::calc_cov_eff_t(sigma_eps,w_full_cov_inv );;
+        Eigen::VectorXd oT_update_mean =    oT_update_cov
                                             *post::calc_mean_eff_T((this->y)[this->T -1], (this->X)[this->T -1], w_full_cov_inv, o_store[T-1],
                                                                beta, sigma_eps, rho);
-        Eigen::EigenMultivariateNormal<float> ot_sampler(oT_update_mean, oT_update_cov);
+        Eigen::EigenMultivariateNormal<double> ot_sampler(oT_update_mean, oT_update_cov);
         o_store[this->T-1] = ot_sampler.samples(1);
 
 
@@ -135,23 +134,23 @@ void ar_model::sample() const {
         //beta update
         vector::vector* beta_vector= beta_stream.add_vec_t();
         beta_vector->set_iter(i);
-        Eigen::MatrixXf beta_update_cov = post::calc_cov_beta((this->X),w_full_cov_inv, this->beta_sig_prior);
+        Eigen::MatrixXd beta_update_cov = post::calc_cov_beta((this->X),w_full_cov_inv, this->beta_sig_prior);
 
-        Eigen::VectorXf beta_update_mean = beta_update_cov * post::calc_mean_beta((this->X), w_full_cov_inv, o_store, rho);
+        Eigen::VectorXd beta_update_mean = beta_update_cov * post::calc_mean_beta((this->X), w_full_cov_inv, o_store, rho);
 
-        Eigen::EigenMultivariateNormal<float> beta_sampler(beta_update_mean, beta_update_cov);
+        Eigen::EigenMultivariateNormal<double> beta_sampler(beta_update_mean, beta_update_cov);
         beta = beta_sampler.samples(1);
 
         //proto beta
-        std::vector<float> beta_vec(beta.data(), beta.data() + beta.rows() * beta.cols());
+        std::vector<double> beta_vec(beta.data(), beta.data() + beta.rows() * beta.cols());
         beta_vector->mutable_vec_value()->Add(beta_vec.begin(), beta_vec.end());
 
         //rho update
         scalar::scalar* rho_scalar= rho_stream.add_scalar();
         rho_scalar->set_iter(i);
 
-        float rho_update_cov = post::calc_var_rho(o_store, w_full_cov_inv, this->rho_sig_prior);
-        float rho_update_mean = rho_update_cov * post::calc_mean_rho(this->X, w_full_cov_inv, o_store, beta);
+        double rho_update_cov = post::calc_var_rho(o_store, w_full_cov_inv, this->rho_sig_prior);
+        double rho_update_mean = rho_update_cov * post::calc_mean_rho(this->X, w_full_cov_inv, o_store, beta);
         std::normal_distribution<double> rho_sampler(rho_update_mean,rho_update_cov);
         rho = rho_sampler(generator);
